+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsUnassignedNumberTreatment/MSFT_TeamsUnassignedNumberTreatment.psm1
@@ -22,7 +22,7 @@ function Get-TargetResource
 
         [Parameter()]
         [System.String]
-        [ValidateSet("User", "ResourceAccount", "Announcement")]
+        [ValidateSet('User', 'ResourceAccount', 'Announcement')]
         $TargetType,
 
         [Parameter()]
@@ -127,7 +127,7 @@ function Set-TargetResource
 
         [Parameter()]
         [System.String]
-        [ValidateSet("User", "ResourceAccount", "Announcement")]
+        [ValidateSet('User', 'ResourceAccount', 'Announcement')]
         $TargetType,
 
         [Parameter()]
@@ -251,7 +251,7 @@ function Test-TargetResource
 
         [Parameter()]
         [System.String]
-        [ValidateSet("User", "ResourceAccount", "Announcement")]
+        [ValidateSet('User', 'ResourceAccount', 'Announcement')]
         $TargetType,
 
         [Parameter()]
@@ -358,7 +358,7 @@ function Export-TargetResource
         $ManagedIdentity
     )
 
-   $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftTeams' `
+    $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftTeams' `
         -InboundParameters $PSBoundParameters
 
     #Ensure the proper dependencies are installed in the current environment.
@@ -396,11 +396,11 @@ function Export-TargetResource
             }
             Write-Host "    |---[$i/$($getValue.Count)] $displayedKey" -NoNewline
             $params = @{
-                Identity = $config.Identity
-                Ensure = 'Present'
-                Credential = $Credential
-                ApplicationId = $ApplicationId
-                TenantId = $TenantId
+                Identity              = $config.Identity
+                Ensure                = 'Present'
+                Credential            = $Credential
+                ApplicationId         = $ApplicationId
+                TenantId              = $TenantId
                 CertificateThumbprint = $CertificateThumbprint
 
             }